# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "api/paulo/user": {
            "post": {
                "tags": [
                    "Paulo"
                ],
                "summary": "Criar um novo usuário",
                "description": "Cria um novo usuário com as informações fornecidas",
                "operationId": "d8821c9c03af24eb9cd8f9004e5e9075",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone_number",
                                    "birth_date",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paulo Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "paulo@exemplo.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "(11) 98765-4321"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-05-23"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "schemas/PauloUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "api/paulo/user/{id}": {
            "get": {
                "tags": [
                    "Paulo"
                ],
                "summary": "Buscar usuário por ID",
                "description": "Retorna os detalhes de um usuário e seus cartões de crédito",
                "operationId": "27751b7de8f4d4720d792b981639a5dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "schemas/PauloUserWithCards"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Paulo"
                ],
                "summary": "Atualizar usuário",
                "description": "Atualiza as informações de um usuário existente",
                "operationId": "d05e772b9d9c2c1a6f304239828c2cd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone_number",
                                    "birth_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paulo Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "paulo@exemplo.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "(11) 98765-4321"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-05-23"
                                    },
                                    "password": {
                                        "description": "Opcional, só enviar se for atualizar",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "schemas/PauloUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Paulo"
                ],
                "summary": "Deletar usuário",
                "description": "Deleta um usuário existente e seus cartões de crédito",
                "operationId": "e5bceaadcf6932a2889728e7df076a46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
...
}